backend/requirements.txt
# Backend dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Vector Database
pinecone-client==3.0.0
chromadb==0.4.22  # For local dev/MVP

# LLM & Agents
langchain==0.1.5
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langgraph==0.0.20

# Task Queue
celery[redis]==5.3.6
redis==5.0.1

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
tiktoken==0.5.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
